revert changes in lockfile package
